+++ src/web/assets/ckeditor/CkeditorAsset.php
@@ -67,11 +67,13 @@ class CkeditorAsset extends BaseCkeditorPackageAsset
                 'Edit {type}',
             ]);
             $view->registerTranslations('ckeditor', [
+                'Add nested content',
                 'Entries cannot be copied between CKEditor fields.',
                 'Entry toolbar',
                 'Entry types list',
                 'Insert link',
                 'Link to the current site',
+                'New {type}',
                 'Site: {name}',
                 'This field doesn’t allow nested entries.',
             ]);